Fix setup.py.
Including the xoutil directory in the Python path confuses setuptools
when installing dependencies that have modules with the same.

The problem is actually global to all of our distributions and should be
fixed in all.

For instance, installing `xotl.ql` requires `xoutil` but since both
include their directories in the PATH and both have a `release` module
the system gets the confused in xoutil and thinks the version it is
installing is the `xotl.ql` version and fails.