Improve interoperability of TimeSpan dates.
Attributes 'start_date' and 'end_date' must return a normal Python
`datetime.date`:class: instance under normal operation.

Returning an Infinity-comparable type should be regarded as an internal
implementation hack to facilitate the implementation of `__and__` and
`__or__`.

This allows normal code to do math with `start_date` and `end_date`.  For
instance::

   >>> from datetime import date
   >>> date.today() - timespan.start_end

That could fail in Python 2.7 because 'date' would not interoperate with the
subtype returned by 'start_date'.