Ensure TimeSpans are coerced to DateTimeSpan for __eq__ and __and__.
All others operators are based on those.

In my test (test_interaction_timespan_with_datetimespans) I didn't even have
to change the implementation of TimeSpan.  But I don't know why Python is
choosing the __and__ DateTimeSpan even when the TimeSpan is the first operand.

So, to be sure, let's our code to deal with it.