DON'T MOVE away from setuptools to poetry/flit based build system.
I'm leaving this commit as witness that I tried to use both Poetry and Flit to
build our library.  The following notes are things that turn me away from the
move.

There are things we loose with this change: versioneer doesn't work with
poetry.  I tried to use poetry-dynamic-versioning without success.  Still I'm
willing to take a peek at Poetry because pipenv is not completely compliant
with PEP-517 (they seem to be working on it).

There are things I dislike of Poetry.  They changed the *standard* versioning
requirements syntax; instead of using, e.g "~=", they use "^".

Trying to make Poetry and Tox work together in our CI pipeline was too time
demanding, so I tried Flit.

But Flit, on the other hand, doesn't even let me name my module 'xotl.tools'.
So it's a no-starter.